finished sprites

+++ h/gpu.h
@@ -24,36 +24,36 @@
 #include<sprite.h>
 
 // check csr ready bit and draw pixel
-extern int ccsrdpx(uint32_t x, uint32_t y, uint16_t color);
+extern int ccsrdpx(uint32_t x, uint32_t y, uint32_t color);
 
 // check csr ready bit and draw line (octant 0)
-extern int ccsrdlnoct0(uint32_t x0, uint32_t y0, uint32_t x1, uint32_t y1, uint16_t color);
+extern int ccsrdlnoct0(uint32_t x0, uint32_t y0, uint32_t x1, uint32_t y1, uint32_t color);
 
 // check csr ready bit and draw line (octant 1)
-extern int ccsrdlnoct1(uint32_t x0, uint32_t y0, uint32_t x1, uint32_t y1, uint16_t color);
+extern int ccsrdlnoct1(uint32_t x0, uint32_t y0, uint32_t x1, uint32_t y1, uint32_t color);
 
 // check csr ready bit and draw line (octant 2)
-extern int ccsrdlnoct2(uint32_t x0, uint32_t y0, uint32_t x1, uint32_t y1, uint16_t color);
+extern int ccsrdlnoct2(uint32_t x0, uint32_t y0, uint32_t x1, uint32_t y1, uint32_t color);
 
 // check csr ready bit and draw line (octant 3)
-extern int ccsrdlnoct3(uint32_t x0, uint32_t y0, uint32_t x1, uint32_t y1, uint16_t color);
+extern int ccsrdlnoct3(uint32_t x0, uint32_t y0, uint32_t x1, uint32_t y1, uint32_t color);
 
 // check csr ready bit draw rectangle
-extern int ccsrdrct(uint32_t x0, uint32_t y0, uint32_t x1, uint32_t y1, uint16_t color);
+extern int ccsrdrct(uint32_t x0, uint32_t y0, uint32_t x1, uint32_t y1, uint32_t color);
 
 // check csr ready bit and draw sprite
-extern int ccsrdspr(uint32_t x, uint32_t y, uint16_t color, const sprite_t* sprite);
+extern int ccsrdspr(uint32_t x, uint32_t y, uint32_t color, const sprite_t* sprite);
 
 // wrapper functions
 
 // check csr ready bit and draw line
-int ccsrdln(uint32_t x0, uint32_t y0, uint32_t x1, uint32_t y1, uint16_t color);
+int ccsrdln(uint32_t x0, uint32_t y0, uint32_t x1, uint32_t y1, uint32_t color);
 
 // check csr ready bit and draw horizontal line
-int ccsrdhln(uint32_t x0, uint32_t x1, uint32_t y, uint16_t color);
+int ccsrdhln(uint32_t x0, uint32_t x1, uint32_t y, uint32_t color);
 
 // check csr ready bit and draw vertical line
-int ccsrdvln(uint32_t x, uint32_t y0, uint32_t y1, uint16_t color);
+int ccsrdvln(uint32_t x, uint32_t y0, uint32_t y1, uint32_t color);
 
 
 #endif//_ARP_GPU_H
